WIP - initial flask refactor; NONFUNCTIONAL
