+++ .gitignore
@@ -1,1 +1,3 @@
 *.pyc
+/venv
+**/__pycache__
